@@ -12,54 +12,54 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.14",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "^7.2.14",
-    "@angular/compiler": "^7.2.14",
-    "@angular/core": "^7.2.14",
-    "@angular/forms": "^7.2.14",
-    "@angular/http": "^7.2.14",
-    "@angular/platform-browser": "^7.2.14",
-    "@angular/platform-browser-dynamic": "^7.2.14",
-    "@angular/pwa": "^0.12.4",
-    "@angular/router": "^7.2.14",
-    "@angular/service-worker": "^7.2.14",
-    "@fortawesome/fontawesome-free": "^5.8.1",
-    "bulma": "0.7.2",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/http": "^7.2.15",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/pwa": "^0.800.0",
+    "@angular/router": "^8.0.0",
+    "@angular/service-worker": "^8.0.0",
+    "@fortawesome/fontawesome-free": "^5.8.2",
+    "bulma": "0.7.5",
+    "core-js": "^3.1.3",
     "localforage": "^1.7.3",
     "ng2-go-top-button": "^7.1.0",
-    "ngforage": "^4.0.1",
-    "ngx-cookie-service": "^2.1.0",
+    "ngforage": "^4.0.3",
+    "ngx-cookie-service": "^2.2.0",
     "ngx-device-detector": "^1.3.5",
     "ngx-pagination": "^3.2.1",
-    "ngx-toastr": "^9.1.1",
-    "rxjs": "^6.5.1",
-    "rxjs-compat": "^6.0.0-rc.0",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.29"
+    "ngx-toastr": "^10.0.4",
+    "rxjs": "^6.5.2",
+    "rxjs-compat": "^6.5.2",
+    "tslib": "^1.9.3",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.1",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "^7.2.14",
-    "@angular/language-service": "^7.2.14",
+    "@angular-devkit/build-angular": "^0.800.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@types/core-js": "2.5.0",
-    "@types/jasmine": "3.3.5",
+    "@types/jasmine": "3.3.13",
     "@types/jasminewd2": "2.0.6",
-    "@types/node": "10.12.18",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "^3.3.0",
+    "@types/node": "12.0.4",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "^3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.1.4",
+    "karma": "^4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.4",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
     "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "^5.4.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~3.2.4"
+    "ts-node": "~8.2.0",
+    "tslint": "~5.16.0",
+    "typescript": "3.4.5"
   },
   "keywords": [
     "angular"
